chore: split the cac parser from the cli
